Update patch set 1
Patch Set 1: Code-Review-1

The problem with this is that if sdnc/core (which builds the sli) is a newer version than sdnc/northbound, you'll end up looking for an older version of the sli.

What we might want to do instead is just remove the version entirely from that <feature> tag.  I don't recall why we added that there, but removing it might be a better change

Patch-set: 1
Label: Code-Review=-1