Update patch set 1
Patch Set 1: Code-Review-1

Please create a new Jira for this item and update the Issue-ID in the commit message with the new Jira. The Jira you used here is closed (it's the Epic for CCVPN in Casablanca - which has been released).  If this is needed for the Casablanca maintenance release, please make sure the severity is either High or Highest and the Fix Version includes Casablanca Maintenance Release.

Patch-set: 1
Label: Code-Review=-1