Proposal to remove OSGi dependencies from the SDNC project
Dependencies on the OSGi frameworks and libraries are removed
by integrating the CCSDK and SDNC project with the lighty.io.
It's a toolkit that allows to use ODL services (in this case
core services and the Restconf) without the dependency
on the Karaf framework and the Blueprint DI.

In this change are created the lighty.io modules which
initialize and expose same services as the Blueprint DI in
the blueprint.xml files.

More info about the lighty.io - https://lighty.io

Signed-off-by: Samuel Kontris <samuel.kontris@pantheon.tech>
Issue-ID: SDNC-803
Change-Id: I0bc0f7c6ed1b152c45b2468bbc57f0f751951461
Signed-off-by: Samuel Kontris <samuel.kontris@pantheon.tech>