Changes to support ODL Sulfur update
Changes to sdnc/northbound to conform to changes in OpenDaylight Sulfur
release.  Many of these changes are due to a change in yangtools, which
no longer implements a common ancestor for its builder classes. This broke
a utility method used in jUnit test cases (build()), which meant all
jUnits using that method had to be updated to call each individual
builder's build() method directly.

Issue-ID: CCSDK-1728
Signed-off-by: Dan Timoney <dtimoney@att.com>
Change-Id: Ia82aa3decd82d873120dd643a6a0427928890f99